Move components to components directory. Remove pages directory
