+++ App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { activateKeepAwake } from 'expo-keep-awake';
 import { UserContextProvider } from './contexts/UserContext';
 import { TrashContextProvider } from './contexts/TrashContext';
-import RootNav from './components/pages/RootNav';
+import RootNav from './components/RootNav';
 
 const App: React.FC = () => {
